MandelbrotPointPosition moved to api package
