+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/api/MandelbrotCalculationProperties.java
@@ -9,8 +9,6 @@ import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import de.lexasoft.mandelbrot.MandelbrotPointPosition;
-
 /**
  * @author nierax
  *
