+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/api/TransitionFactoryTest.java
@@ -19,8 +19,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import de.lexasoft.mandelbrot.MandelbrotPointPosition;
-
 /**
  * @author nierax
  *
